refactor: reorganize imports and update media handling structure
