+++ agentflow_cli/src/app/core/config/settings.py
@@ -2,6 +2,7 @@ import logging
 import os
 from functools import lru_cache
 
+from pydantic import ConfigDict
 from pydantic import field_validator, model_validator
 from pydantic_settings import BaseSettings
 
@@ -149,8 +150,7 @@ class Settings(BaseSettings):
 
         return self
 
-    class Config:
-        extra = "allow"
+    model_config = ConfigDict(extra="allow")
 
 
 @lru_cache
