+++ agentflow_cli/src/app/core/exceptions/handle_errors.py
@@ -1,5 +1,5 @@
 # Handle all exceptions of agentflow here
-from agentflow.exceptions import (
+from agentflow.core.exceptions import (
     GraphError,
     GraphRecursionError,
     MetricsError,
