+++ agentflow_cli/src/app/loader.py
@@ -4,9 +4,9 @@ import logging
 import os
 from pathlib import Path
 
-from agentflow.checkpointer import BaseCheckpointer
-from agentflow.graph import CompiledGraph
-from agentflow.store import BaseStore
+from agentflow.core import CompiledGraph
+from agentflow.storage.checkpointer import BaseCheckpointer
+from agentflow.storage.store import BaseStore
 from injectq import InjectQ
 
 from agentflow_cli import BaseAuth
