+++ agentflow_cli/src/app/routers/checkpointer/services/checkpointer_service.py
@@ -1,7 +1,7 @@
 from typing import Any
 
-from agentflow.checkpointer import BaseCheckpointer
-from agentflow.state import AgentState, Message
+from agentflow.core.state import AgentState, Message
+from agentflow.storage.checkpointer import BaseCheckpointer
 from fastapi import HTTPException
 from injectq import inject, singleton
