+++ agentflow_cli/src/app/routers/graph/schemas/graph_schemas.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from agentflow.state import Message
+from agentflow.core.state import Message
 from agentflow.utils import ResponseGranularity
 from pydantic import BaseModel, Field, field_validator
