+++ agentflow_cli/src/app/routers/graph/services/graph_service.py
@@ -3,9 +3,9 @@ from collections.abc import AsyncIterable
 from typing import Any
 from uuid import uuid4
 
-from agentflow.checkpointer import BaseCheckpointer
-from agentflow.graph import CompiledGraph
-from agentflow.state import AgentState, Message, StreamChunk, StreamEvent
+from agentflow.core.graph import CompiledGraph
+from agentflow.core.state import AgentState, Message, StreamChunk, StreamEvent
+from agentflow.storage.checkpointer import BaseCheckpointer
 from agentflow.utils.thread_info import ThreadInfo
 from fastapi import HTTPException
 from injectq import InjectQ, inject, singleton
