+++ agentflow_cli/src/app/utils/media/__init__.py
@@ -5,12 +5,8 @@ the core library stays lightweight — SDK users extract text themselves.
 The API platform auto-extracts using textxtract.
 """
 
-from ._compat import ensure_document_handling_aliases
 from .extractor import DocumentExtractor
 from .pipeline import DocumentPipeline
-
-
-ensure_document_handling_aliases()
 
 
 __all__ = [
