+++ tests/unit_tests/store/test_store_service.py
@@ -4,7 +4,7 @@ from unittest.mock import AsyncMock
 from uuid import uuid4
 
 import pytest
-from agentflow.store.store_schema import DistanceMetric, MemoryType, RetrievalStrategy
+from agentflow.storage.store.store_schema import DistanceMetric, MemoryType, RetrievalStrategy
 
 from agentflow_cli.src.app.routers.store.schemas.store_schemas import (
     DeleteMemorySchema,
