+++ tests/unit_tests/test_checkpointer_service.py
@@ -3,8 +3,8 @@
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
-from agentflow.checkpointer import BaseCheckpointer
-from agentflow.state import AgentState, Message
+from agentflow.storage.checkpointer import BaseCheckpointer
+from agentflow.core.state import AgentState, Message
 
 from agentflow_cli.src.app.routers.checkpointer.schemas.checkpointer_schemas import (
     MessagesListResponseSchema,
